+++ Changes.md
@@ -17,6 +17,11 @@ Fixes
 
 - PathListingWidget : Fixed parent layout update when column sizes change.
 
+API
+---
+
+- TogglePlugValueWidget : Added new widget type for toggling between default and non-default values.
+
 1.6.0.0 (relative to 1.5.16.2)
 =======